:sparkles: Rewrite ite for bv
