+++ grisette.cabal
@@ -124,7 +124,6 @@ library
     Grisette.Internal.SymPrim.IntBitwidth
     Grisette.Internal.SymPrim.ModelRep
     Grisette.Internal.SymPrim.Prim.Internal.Caches
-    Grisette.Internal.SymPrim.Prim.Internal.Instances.BVPEval
     Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalBitCastTerm
     Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalDivModIntegralTerm
     Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalFloatingTerm
