+++ src/Grisette/Internal/SymPrim/Prim/Term.hs
@@ -12,14 +12,12 @@ module Grisette.Internal.SymPrim.Prim.Term
   ( module Grisette.Internal.SymPrim.Prim.Internal.Term,
     module Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalShiftTerm,
     module Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalRotateTerm,
-    module Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalNumTerm,
     module Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalOrdTerm,
     module Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalDivModIntegralTerm,
     module Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalIEEEFPConvertibleTerm,
   )
 where
 
-import Grisette.Internal.SymPrim.Prim.Internal.Instances.BVPEval ()
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalBitCastTerm ()
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalDivModIntegralTerm
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalFP ()
@@ -27,7 +25,7 @@ import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalFloatingTerm ()
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalFractionalTerm ()
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalFromIntegralTerm ()
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalIEEEFPConvertibleTerm
-import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalNumTerm
+import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalNumTerm ()
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalOrdTerm
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalRotateTerm
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalShiftTerm
